// src/constants/enums.ts
export enum Principle {
  Hierarchy = 'Hierarchy',
  Alignment = 'Alignment',
  Unity = 'Unity',
  Balance = 'Balance',
  Readability = 'Readability',
}

export enum ViolationType {
  H_1 = 'H_1',
  H_2 = 'H_2',
  H_3 = 'H_3',
  A_1 = 'A_1',
  A_2 = 'A_2',
  B_1 = 'B_1',
  B_2 = 'B_2',
  B_3 = 'B_3',
  U_1 = 'U_1',
  U_2 = 'U_2',
  U_3 = 'U_3',
  U_4 = 'U_4',
  R_1 = 'R_1',
  R_2 = 'R_2',
  R_3 = 'R_3',
  None = 'None',
}

export enum Condition {
  C = 'C', // Realtime
}

export enum LogType {
  ResolveIssue = 'ResolveIssue',
  DismissIssue = 'DismissIssue',
  ExpandIssueCard = 'ExpandIssueCard',
  RequestFeedback = 'RequestFeedback',
  TurnOnAnnotation = 'TurnOnAnnotation',
  TurnOffAnnotation = 'TurnOffAnnotation',
  ClickCommon = 'ClickCommon',
  ClickDetected = 'ClickDetected',
  ClickPrincipleTab = 'ClickPrincipleTab',
  ClickPrincipleReadMore = 'ClickPrincipleReadMore',
  MoveToNext = 'MoveToNext',
  Start = 'Start',
  End = 'End',
  CreateIssue = 'CreateIssue',
  DeleteIssue = 'DeleteIssue',
  PublishIssue = 'PublishIssue',
  UnpublishIssue = 'UnpublishIssue',
}

export enum LinearDirection {
  Horizontal = 'Horizontal',
  Vertical = 'Vertical',
}

export enum LinearJustify {
  Center = 'Center',
  Start = 'Start',
  End = 'End',
  SpaceBetween = 'SpaceBetween',
  SpaceAround = 'SpaceAround',
}

export enum LinearAlign {
  Center = 'Center',
  Start = 'Start',
  End = 'End',
  Stretch = 'Stretch',
}

export enum FontList {
  EB_Garamond = 'EB Garamond',
  Inria_Serif = 'Inria Serif',
  IBM_Plex_Serif = 'IBM Plex Serif',
  Nixie_One = 'Nixie One',
  Elsie = 'Elsie',
  Open_Sans = 'Open Sans',
  IBM_lex_Sans = 'IBM Plex Sans',
  Noto_Sans = 'Noto Sans',
  PT_Sans_Narrow = 'PT Sans Narrow',
  Archivo_Narrow = 'Archivo Narrow',
  Caveat = 'Caveat',
  Galada = 'Galada',
  Permanent_Marker = 'Permanent Marker',
  Rock_Salt = 'Rock Salt',
  Oleo_Script = 'Oleo Script',
  Patrick_Hand = 'Patrick Hand',
  IBM_Plex_Mono = 'IBM Plex Mono',
  Roboto_Mono = 'Roboto Mono',
  VT323 = 'VT323',
  Ubuntu_Mono = 'Ubuntu Mono',
  Cutive_Mono = 'Cutive Mono',
  Abril_Fatface = 'Abril Fatface',
  Bangers = 'Bangers',
  Fredoka_One = 'Fredoka One',
  Graduate = 'Graduate',
  Koulen = 'Koulen',
  Bungee = 'Bungee',
}